Refac: rename file_io to image_io
