+++ src/careamics/compat/careamist.py
@@ -34,7 +34,7 @@ from careamics.config.support import (
     SupportedData,
     SupportedLogger,
 )
-from careamics.file_io import WriteFunc, get_write_func
+from careamics.image_io import WriteFunc, get_write_func
 from careamics.lightning.data.data_module_utils import list_files
 from careamics.lightning.utils.csv_logger import read_csv_logger
 from careamics.utils import get_logger
