+++ src/careamics/compat/dataset/dataset_utils/iterate_over_files.py
@@ -10,7 +10,7 @@ from numpy.typing import NDArray
 from torch.utils.data import get_worker_info
 
 from careamics.compat.config.data import DataConfig, InferenceConfig
-from careamics.file_io.read import read_tiff
+from careamics.image_io.read import read_tiff
 from careamics.utils.logging import get_logger
 from careamics.utils.reshape_array import reshape_array
