+++ src/careamics/compat/dataset/iterable_dataset.py
@@ -13,7 +13,7 @@ from careamics.compat.config.augmentations import NormalizeConfig
 from careamics.compat.config.data import DataConfig
 from careamics.compat.transforms.compose import Compose
 from careamics.dataset.normalization.running_mean_std import WelfordStatistics
-from careamics.file_io.read import read_tiff
+from careamics.image_io.read import read_tiff
 from careamics.utils.logging import get_logger
 
 from .dataset_utils import iterate_over_files
