+++ src/careamics/compat/dataset/iterable_pred_dataset.py
@@ -12,7 +12,7 @@ from torch.utils.data import IterableDataset
 from careamics.compat.config.augmentations import NormalizeConfig
 from careamics.compat.config.data import InferenceConfig
 from careamics.compat.transforms.compose import Compose
-from careamics.file_io.read import read_tiff
+from careamics.image_io.read import read_tiff
 
 from .dataset_utils import iterate_over_files
