+++ src/careamics/compat/lightning/callbacks/prediction_writer_callback/write_strategy.py
@@ -13,7 +13,7 @@ from careamics.compat.config.data.tile_information import TileInformation
 from careamics.compat.dataset import IterablePredDataset, IterableTiledPredDataset
 from careamics.compat.dataset.dataset_utils.file_utils import create_write_file_path
 from careamics.compat.prediction_utils.stitch_prediction import stitch_prediction_single
-from careamics.file_io import WriteFunc
+from careamics.image_io import WriteFunc
 from careamics.utils import get_logger
 
 from .file_path_utils import get_sample_file_path
