+++ src/careamics/compat/lightning/predict_data_module.py
@@ -18,7 +18,7 @@ from careamics.compat.dataset import (
 )
 from careamics.compat.dataset.tiling.collate_tiles import collate_tiles
 from careamics.config.support import SupportedData
-from careamics.file_io.read import get_read_func
+from careamics.image_io.read import get_read_func
 from careamics.lightning.data.data_module_utils import list_files
 from careamics.utils import get_logger
