+++ src/careamics/dataset/image_stack/file_image_stack.py
@@ -8,7 +8,7 @@ import numpy as np
 import tifffile
 from numpy.typing import DTypeLike, NDArray
 
-from careamics.file_io.read import ReadFunc, read_tiff
+from careamics.image_io.read import ReadFunc, read_tiff
 from careamics.utils.reshape_array import AxesTransform, reshape_array
 
 from .image_utils.image_stack_utils import channel_slice, pad_patch
