+++ src/careamics/image_io/read/get_func.py
