+++ src/careamics/image_io/read/tiff.py
