+++ src/careamics/image_io/write/__init__.py
