+++ src/careamics/image_io/write/get_func.py
