+++ src/careamics/image_io/write/tiff.py
