+++ tests/compat/lightning/callbacks/prediction_writer_callback/test_old_cache_tiles_write_strategy.py
@@ -15,7 +15,7 @@ from careamics.compat.dataset.tiling import extract_tiles
 from careamics.compat.lightning.callbacks.prediction_writer_callback import (
     CacheTiles,
 )
-from careamics.file_io import WriteFunc
+from careamics.image_io import WriteFunc
 
 
 def create_tiles(n_samples: int) -> tuple[list[NDArray], list[TileInformation]]:
