+++ tests/file_io/read/test_get_read_func.py
@@ -1,8 +1,8 @@
 import pytest
 
 from careamics.config.support import SupportedData
-from careamics.file_io import get_read_func
-from careamics.file_io.read import read_tiff
+from careamics.image_io import get_read_func
+from careamics.image_io.read import read_tiff
 
 
 def test_get_read_tiff():
