+++ tests/file_io/read/test_read_tiff.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import tifffile
 
-from careamics.file_io.read import read_tiff
+from careamics.image_io.read import read_tiff
 
 
 def test_read_tiff(tmp_path, ordered_array):
