+++ tests/file_io/write/test_get_write_func.py
@@ -1,6 +1,6 @@
 from careamics.config.support import SupportedData
-from careamics.file_io import get_write_func
-from careamics.file_io.write import write_tiff
+from careamics.image_io import get_write_func
+from careamics.image_io.write import write_tiff
 
 
 def test_get_write_tiff():
